v1.1.2

+++ pkg/cloudapi/account/models.go
@@ -78,10 +78,10 @@ type Resource struct {
 	CPU int64 `json:"cpu"`
 
 	// Disk size
-	DiskSize int64 `json:"disksize"`
+	DiskSize float64 `json:"disksize"`
 
 	// Max disk size
-	DiskSizeMax int64 `json:"disksizemax"`
+	DiskSizeMax uint64 `json:"disksizemax"`
 
 	// Number of External IPs
 	ExtIPs int64 `json:"extips"`
@@ -105,7 +105,7 @@ type DiskUsage struct {
 	DiskSize float64 `json:"disksize"`
 
 	// Disk size max
-	DiskSizeMax float64 `json:"disksizemax"`
+	DiskSizeMax uint64 `json:"disksizemax"`
 }
 
 // Information about resources
